removed RAExpressionAttributeOccurrencesOperations
